Properly respect maxseenage when registering offloads for burst detection
Just checking that the ip:file is in the hash is not enough, also need
to check the actual time of the entry :-)

While at it, avoid hard-coded value for cutoff increase when looping to reduce
hash size below maxseenentries in size.

Also, don't use the unrelated purgeinterval config setting for the
ip:file hash, use something based on maxseenage instead.