Split in-memory and DBM caching
We see a lot of workload that sweeps the entire site with a HEAD+GET on
each file. This doesn't play too well with caching everything into a
DBM file.

Split caching into cache (in-memory) and db (DBM file).
Provide config knobs to tune how many accesses within a given
time frame that are required for an entry to be promoted into the DBM.

Add required functionality to purge them both down to size according to
the set limits.

Config file changes:

Split maxentries into maxcache and maxdb, for the number of in-memory
and in-DBM max sizes respectively. maxcache must be at least as large as
maxdb.

Added maxhitage and hitspromote. After hitspromote hits, each at most
maxhitage seconds apart, an entry is promoted from being cached in memory
to being cached in the DBM.