redirprg: Implement config-file validation
Previously we relied on the config file being 100% correct and just
used the config object as is. This caused very weird errors if something
was spelled wrong, for example.

Now we do some rudimentary checking that there is only stuff that we
understand in the config file, and provide default values and
explanation in the code of all config items.