redirprg: Split DB purging in multiple stages to not block lookups
Split DB purging into a find stage and a deletion stage, where the
deletion stage pauses after timeout to allow any pending lookups
to be processed.

Additionally, group DB and cache entries to delete and only tie the
DB for writing when needed.

This should reduce the long processing pauses when massive amounts of
entries are purged from the DB.